[AUTO-68] chore: refactor pages components to templates
