--- src/components/pages/index.ts
@@ -1,4 +1,0 @@
-export { default as DashPageError } from './dash-page-error';
-export { default as DashPageLoading } from './dash-page-loading';
-export { default as DashPageUILayout } from './dash-page-ui-layout';
-export { default as Contract } from './contract';
