+++ src/components/sections/contract/contract.styles.ts
@@ -1,0 +1,135 @@
+import { StyleSheet } from '@react-pdf/renderer';
+
+export const styles = StyleSheet.create({
+  page: {
+    display: 'flex',
+    flexDirection: 'column',
+    backgroundColor: '#fff',
+    color: '#000',
+    fontSize: '16px',
+  },
+  pageTitle: {
+    textAlign: 'center',
+    fontWeight: 'bold',
+    fontSize: '24px',
+  },
+  flexColCentered: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: '100%',
+  },
+  titleSection: {
+    display: 'flex',
+    width: '100%',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    textAlign: 'center',
+  },
+  dateTimeSection: {
+    display: 'flex',
+    width: '100%',
+    paddingTop: '40px',
+  },
+  partiesSection: {
+    display: 'flex',
+    width: '100%',
+    flexDirection: 'column',
+    gap: '8px',
+    paddingTop: '40px',
+  },
+  largeTitle: {
+    fontSize: '20px',
+    fontWeight: 'bold',
+  },
+  largeTitleCentered: {
+    fontSize: '20px',
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginTop: '40px',
+  },
+  schoolPartySection: {
+    display: 'flex',
+    width: '100%',
+    flexDirection: 'column',
+    paddingLeft: '12px',
+    paddingRight: '12px',
+    paddingTop: '12px',
+  },
+  studentPartySection: {
+    display: 'flex',
+    width: '100%',
+    flexDirection: 'column',
+    paddingLeft: '12px',
+    paddingRight: '12px',
+  },
+  largeText: {
+    fontSize: '20px',
+    marginBottom: '6px',
+  },
+  largeTextBold: {
+    fontSize: '20px',
+    fontWeight: 'bold',
+  },
+  informationContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    width: '100%',
+    paddingLeft: '12px',
+    paddingRight: '12px',
+    gap: '4px',
+  },
+  flexFullWidth: {
+    display: 'flex',
+    width: '100%',
+  },
+  flex1: {
+    flex: 1,
+  },
+  flex1Centered: {
+    flex: 1,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    textAlign: 'center',
+  },
+  flex1CenteredBold: {
+    flex: 1,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    textAlign: 'center',
+    fontWeight: 'bold',
+  },
+  conditionsContainer: {
+    display: 'flex',
+    width: '100%',
+    flexDirection: 'column',
+    gap: '16px',
+    paddingTop: '16px',
+  },
+  condition: {
+    display: 'flex',
+    flexDirection: 'column',
+    width: '100%',
+    paddingLeft: '12px',
+    paddingRight: '12px',
+  },
+  flexWithGap: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: '16px',
+    width: '100%',
+    marginBottom: '120px',
+  },
+  flexCenteredText: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    textAlign: 'center',
+    marginTop: '40px',
+    marginBottom: '40px',
+  },
+});
