+++ src/components/sections/contract/contract.tsx
@@ -1,25 +1,10 @@
 'use client';
 
-import { Page, Text, View, Document, StyleSheet } from '@react-pdf/renderer';
+import { Page, Text, View, Document } from '@react-pdf/renderer';
 import { useEffect } from 'react';
 
-export interface ContractProps {
-  date: string;
-  city: string;
-  licenseFile: { id: string; category: string; practiceHours: string; theoryHours: string };
-  school: {
-    name: string;
-    numRegistrePermis: string;
-    address: string;
-    numRegistreFiscale: string;
-    numRegistreCommerce: string;
-    city: string;
-    fax: string;
-    phone: string;
-    email: string;
-  };
-  student: { cin: string; birthdate: string; birthplace: string; address: string; numReferenceWeb: string };
-}
+import type { ContractProps } from './contract.types';
+import { styles } from './contract.styles';
 
 export default function Contract({
   date,
@@ -218,137 +203,3 @@ export default function Contract({
     </main>
   );
 }
-
-const styles = StyleSheet.create({
-  page: {
-    display: 'flex',
-    flexDirection: 'column',
-    backgroundColor: '#fff',
-    color: '#000',
-    fontSize: '16px',
-  },
-  pageTitle: {
-    textAlign: 'center',
-    fontWeight: 'bold',
-    fontSize: '24px',
-  },
-  flexColCentered: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: '100%',
-  },
-  titleSection: {
-    display: 'flex',
-    width: '100%',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    textAlign: 'center',
-  },
-  dateTimeSection: {
-    display: 'flex',
-    width: '100%',
-    paddingTop: '40px',
-  },
-  partiesSection: {
-    display: 'flex',
-    width: '100%',
-    flexDirection: 'column',
-    gap: '8px',
-    paddingTop: '40px',
-  },
-  largeTitle: {
-    fontSize: '20px',
-    fontWeight: 'bold',
-  },
-  largeTitleCentered: {
-    fontSize: '20px',
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginTop: '40px',
-  },
-  schoolPartySection: {
-    display: 'flex',
-    width: '100%',
-    flexDirection: 'column',
-    paddingLeft: '12px',
-    paddingRight: '12px',
-    paddingTop: '12px',
-  },
-  studentPartySection: {
-    display: 'flex',
-    width: '100%',
-    flexDirection: 'column',
-    paddingLeft: '12px',
-    paddingRight: '12px',
-  },
-  largeText: {
-    fontSize: '20px',
-    marginBottom: '6px',
-  },
-  largeTextBold: {
-    fontSize: '20px',
-    fontWeight: 'bold',
-  },
-  informationContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    width: '100%',
-    paddingLeft: '12px',
-    paddingRight: '12px',
-    gap: '4px',
-  },
-  flexFullWidth: {
-    display: 'flex',
-    width: '100%',
-  },
-  flex1: {
-    flex: 1,
-  },
-  flex1Centered: {
-    flex: 1,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    textAlign: 'center',
-  },
-  flex1CenteredBold: {
-    flex: 1,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    textAlign: 'center',
-    fontWeight: 'bold',
-  },
-  conditionsContainer: {
-    display: 'flex',
-    width: '100%',
-    flexDirection: 'column',
-    gap: '16px',
-    paddingTop: '16px',
-  },
-  condition: {
-    display: 'flex',
-    flexDirection: 'column',
-    width: '100%',
-    paddingLeft: '12px',
-    paddingRight: '12px',
-  },
-  flexWithGap: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '16px',
-    width: '100%',
-    marginBottom: '120px',
-  },
-  flexCenteredText: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    textAlign: 'center',
-    marginTop: '40px',
-    marginBottom: '40px',
-  },
-});
