+++ src/components/sections/contract/index.ts
@@ -1,0 +1,1 @@
+export { default as Contract } from './contract';
