+++ src/components/templates/dashboard/error/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardError } from './dashboard-error';
