+++ src/components/templates/dashboard/layout/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardLayout } from './dashboard-layout';
