+++ src/components/templates/dashboard/loading/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardLoading } from './dashboard-loading';
